quiz에서 scoreboard용 API 분리
